create the ownerId field on Products table to control who can delete or edit a product

+++ src/use-cases/delete-product.ts
@@ -1,18 +1,27 @@
 import { ProductsRepository } from '@/database/repositories/products-repository'
 import { ResourceNotFoundError } from './errors/resource-not-found-error'
+import { UUID } from 'node:crypto'
+import { NotAllowedError } from './errors/not-allowed-error'
 
 interface DeleteProductUseCaseRequest {
-  productId: string
+  productId: UUID
+  ownerId: UUID
 }
 
 export class DeleteProductUseCase {
   constructor(private productsRepository: ProductsRepository) {}
 
-  async execute({ productId }: DeleteProductUseCaseRequest) {
+  async execute({ productId, ownerId }: DeleteProductUseCaseRequest) {
     const product = await this.productsRepository.findById(productId)
 
     if (!product) {
       throw new ResourceNotFoundError()
+    }
+
+    const isOwner = await this.productsRepository.verifyProductOwner(productId, ownerId)
+
+    if (!isOwner) {
+      throw new NotAllowedError()
     }
 
     await this.productsRepository.delete(product)
